+++ R/gh_token.R
@@ -17,37 +17,53 @@
 #' @section PATs for GitHub Enterprise:
 #'
 #' gh lets you use different PATs for different GitHub API URLs, by looking
-#' for the PAT in an URL specific environment variable first. It uses
-#' [slugify_url()] to compute a suffix from the API URL, by extracting the
-#' host name and removing the protocol and the path from it, and replacing
-#' special characters with underscores. This suffix is added to
-#' `GITHUB_PAT_` then. For example for the default API URL:
-#' <https://api.github.com>, the `GITHUB_PAT_API_GITHUB_COM` environment
-#' variable is consulted first.
+#' for the PAT in an URL-specific environment variable first. The helper
+#' [slugify_url()] computes a suffix from the API URL like so:
+#' * Extract the host name, i.e. drop both the protocol and any path
+#' * Substitute "github.com" for "api.github.com"
+#' * Replace special characters with underscores
+#' * Convert to ALL CAPS
 #'
-#' You can set the default API URL via the `GITHUB_API_URL` environment
+#' This suffix is then added to `GITHUB_PAT_` to form the name of an environment
+#' variable. It's probably easiest to just look at some examples.
+#'
+#' ```{r}
+#' # both give same result
+#' slugify_url("https://api.github.com")
+#' slugify_url("https://github.com")
+#'
+#' # an instance of GitHub Enterprise
+#' # both give same result
+#' slugify_url("https://github.acme.com")
+#' slugify_url("https://github.acme.com/api/v3")
+#' ```
+#'
+#' This implies that, for the default API URL <https://api.github.com>, these
+#' env vars are consulted, in this order:
+#' * `GITHUB_PAT_GITHUB_COM`
+#' * `GTIHUB_PAT`
+#' * `GITHUB_TOKEN`
+#'
+#' You can customize the default API URL via the `GITHUB_API_URL` environment
 #' variable.
-#'
-#' If the API URL specific environment variable is not set, then gh falls
-#' back to `GITHUB_PAT` and then to `GITHUB_TOKEN`.
 #'
 #' @section Storing PATs in the system keyring:
 #'
 #' gh supports storing your PAT in the system keyring, on Windows, macOS
 #' and Linux, using the keyring package. To turn on keyring support, you
-#' need to set the `GH_KEYRING` environment variables to `true`, in your
-#' `.Renviron` file or profile.
+#' need to set the `GH_KEYRING` environment variables to `true`, e.g. in your
+#' `.Renviron` file.
 #'
 #' If keyring support is turned on, then for each PAT environment variable,
 #' gh first checks for it via `Sys.getenv()` and, if unset, gh then checks
 #' whether such a key exists in the system keyring and, if yes, it uses
 #' the associated value as the PAT. I.e. without a custom `GITHUB_API_URL`
-#' variable, gh checks the `GITHUB_PAT_API_GITHUB_COM` env var first, then
+#' variable, gh checks the `GITHUB_PAT_GITHUB_COM` env var first, then
 #' checks for that key in the keyring, then moves on to do same with
-#' `GITHUB_PAT`, etc. The keyring check looks like this:
+#' `GITHUB_PAT` and perhaps `GITHUB_TOKEN`. The keyring check looks like this:
 #'
 #' ```r
-#' keyring::key_get("GITHUB_PAT_API_GITHUB_COM")
+#' keyring::key_get("GITHUB_PAT_GITHUB_COM")
 #' ```
 #'
 #' and it uses the default keyring backend and the default keyring within
@@ -166,35 +182,67 @@ gh_auth <- function(token) {
 
 #' Compute the suffix that gh uses for GitHub API URL specific PATs
 #'
-#' @param url Character vector HTTP/HTTPS URLs.
+#' `slugify_url()` determines a suffix from a URL and this suffix is used to
+#' construct the name of an environment variable that holds the PAT for a
+#' specific GitHub URL. This is mostly relevant to people using GitHub
+#' Enterprise.
+#'
+#' @param url Character vector of HTTP/HTTPS URLs. They don't have to be in the
+#'   API-specific form, although they can be.
 #' @return Character vector of suffixes.
 #'
 #' @seealso [gh_token()]
 #' @export
 #' @examples
-#' # The main GH site
+#' # main github.com site
 #' slugify_url("https://api.github.com")
+#' slugify_url("https://github.com")
 #'
-#' # A custom one
+#' # an instance of GitHub Enterprise
 #' slugify_url("https://github.acme.com")
+#' slugify_url("https://github.acme.com/api/v3")
+slugify_url <- function(url) {           # https://jane@github.uni.edu/api/v3
+  url <- get_baseurl(url)                # https://jane@github.uni.edu
+  url <- normalize_host(url)
+  x2 <- sub("^.*://([^/]*@)?", "", url)  # github.uni.edu
+  x3 <- gsub("[.]+", "_", x2)            # github_uni_edu
+  x4 <- gsub("[^-a-zA-Z0-9_]", "", x3)
+  toupper(x4)                            # GITHUB_UNI_EDU
+}
 
-slugify_url <- function(url) {
+get_baseurl <- function(url) {               # https://github.uni.edu/api/v3/
   if (!any(grepl("^https?://", url))) {
     stop("Only works with HTTP(S) protocols")
   }
-  x2 <- sub("^.*://([^/]*@)?", "", url)
-  x3 <- sub("/+$", "", x2)
-  x4 <- gsub("[./]+", "_", x3)
-  x5 <- gsub("[^-a-zA-Z0-9_]", "", x4)
-  toupper(x5)
+  prot <- sub("^(https?://).*$", "\\1", url) # https://
+  rest <- sub("^https?://(.*)$", "\\1", url) #         github.uni.edu/api/v3/
+  host <- sub("/.*$", "", rest)              #         github.uni.edu
+  paste0(prot, host)                         # https://github.uni.edu
 }
 
-get_baseurl <- function(x) {
-  if (!any(grepl("^https?://", x))) stop("Only works with HTTP(S) protocols")
-  prot <- sub("^(https?://).*$", "\\1", x)
-  rest <- sub("^https?://(.*)$", "\\1", x)
-  host <- sub("/.*$", "", rest)
-  paste0(prot, host)
+normalize_host <- function(x) {
+  sub("api[.]github[.]com", "github.com", x)
+}
+
+get_hosturl <- function(url) {
+  url <- get_baseurl(url)
+  normalize_host(url)
+}
+
+get_apiurl <- function(url) {
+  host_url <- get_hosturl(url)
+  prot_host <- strsplit(host_url, "://", fixed = TRUE)[[1]]
+  if (is_github_dot_com(host_url)) {
+    paste0(prot_host[[1]], "://api.github.com")
+  } else {
+    paste0(host_url, "/api/v3")
+  }
+}
+
+is_github_dot_com <- function(url) {
+  url <- get_baseurl(url)
+  url <- normalize_host(url)
+  grepl("^https?://github.com", url)
 }
 
 # gh_pat class: exists in order have a print method that hides info ----